+++ src/screen/password-redefine/redefinirSenha.tsx
@@ -6,16 +6,15 @@ import {
   TitleRedefinePass,
 } from "./redefinirSenha-style";
 import logo from "../../assets/images/logo/logo.png";
-
-import Header from "../../components/Headers/header/header";
-import FooterCurto from "../../components/Footers/footer/footer";
 import { useNavigation } from "@react-navigation/native";
 import { ContainerSafe, ImageIconLogo, Scroll } from "../login/login-style";
 import { ContainerTelaPrincipal } from "../home/telaPrincipal-style";
 import InputGeral from "../../components/Forms/inputGeral/inputGeral";
 import { Platform } from "react-native";
 import { propsStack } from "@/src/routes/Stack/Models";
-import Button from "@/src/components/Buttons/button/button";
+import Button from "@/src/components/Buttons/button/button"
+import Footer from "../../components/Footers/footerCorrect/footerCorrect";
+import Header from "../../components/Footers/footerCorrect/headerCorrect";;
 
 const RedefinirSenha = () => {
   const navigation = useNavigation<propsStack>();
@@ -24,7 +23,7 @@ const RedefinirSenha = () => {
     <ContainerSafe behavior={Platform.OS === "ios" ? "padding" : "height"}>
       <Scroll>
         <ContainerTelaPrincipal>
-          <Header back={true} />
+          <Header  />
           <ImageIconLogoRedefinir source={logo} />
           <ContainerEnviaremos>
             <TitleRedefinePass>Redefinir senha</TitleRedefinePass>
@@ -33,7 +32,7 @@ const RedefinirSenha = () => {
 
           <InputGeral inputtext={'Email'} />
           <Button buttontext={'Avançar'} />
-          <FooterCurto />
+          <Footer />
         </ContainerTelaPrincipal>
       </Scroll>
     </ContainerSafe>